code restructed.
